big refactors / add prison properties / cell binding
Add prisoner properties to check whether they are stripped naked / to underwear.
Add prisoner properties to determine what type of jail cell they must go to, random, empty, or gender exclusive.
Fix prisoner properties when the effect is no longer active (deconstructed through OnEffectFinish), by storing their value in ArrestVars instead of locally in the script.
Add Prison properties to determine the jail cell assigning priority, and what type of jail cells are to be assigned, either empty or gender exclusive, or random if none is set.
Add a way to bind the jail cell to the prisoner and in the process bind the cell door to the jail cell.
Add multiple options to Jail Cell script