fix jail cell prisoners not updating / additional Prison methods
Fix JailCell prisoners map not updating, they would reset when another prison entered the cell that was already initialized with at least one prisoner (The map was not being retained in memory, so the value would get destroyed when the function went out of scope).
Remove "this" pointer for JailCells, considering they are already identified through "self".

Add Prison.GetJailCellBasedOnPriority() to retrieve a jail cell based on the Prison's config and priorities.